requirements.txt
simple-term-menu
injector
python-magic-bin==0.4.14
transitions